% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_colors.R
\name{au_colors}
\alias{au_colors}
\title{Aarhus University brand colors}
\usage{
au_colors(...)
}
\arguments{
\item{...}{Name(s) of color(s) to be used. Valid color names: \code{blue}, \code{darkblue},
\code{purple}, \code{darkpurple}, \code{cyan}, \code{darkcyan}, \code{turkis}, \code{darkturkis}, \code{green},
\code{darkgreen}, \code{yellow}, \code{darkyellow}, \code{orange}, \code{darkorange}, \code{red}, \code{darkred},
\code{magenta}, \code{darkmagenta}, \code{grey}, \code{darkgrey}, \code{white}.}
}
\value{
A named vector with the chosen colors and their corresponding hex codes.
}
\description{
Helper function to retrieve an arbitrary number of AU brand colors by name.
Returns a named vector containing the color name(s) as passed into the function,
and the corresponding AU hex code(s).
}
\details{
Implementation is based on the \href{https://medarbejdere.au.dk/en/administration/communication/guidelines/guidelinesforcolours}{official brand guidelines}.
}
\examples{
au_colors("red", "blue")
}
