% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_au.R
\name{scale_fill_au}
\alias{scale_fill_au}
\title{Title}
\usage{
scale_fill_au(
  aesthetics = "fill",
  style = "light",
  reverse = FALSE,
  discrete = F,
  colors = NA,
  ...,
  breaks = ggplot2::waiver(),
  na.value = au_colors("grey")
)
}
\description{
Title
}
