% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplotter.R
\name{violinplotter}
\alias{violinplotter}
\title{Title}
\usage{
violinplotter(
  data,
  x_val,
  y_val,
  x_lab = "x",
  y_lab = "y",
  title = "",
  filter = F,
  filter_col = NA,
  filter_val = NA,
  comp_vec = NA,
  col_vec = au_colors()
)
}
\description{
Title
}
