% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplotter.R
\name{volcanoplotter}
\alias{volcanoplotter}
\title{Volcanoplotter}
\usage{
volcanoplotter(
  data,
  x_val,
  y_val,
  nonsig_col = "grey",
  pointsize = 1,
  point_alpha = 1,
  y_lab = "P-value",
  x_lab = "logFC",
  title = "",
  sig_neg_col = au_colors("blue"),
  sig_pos_col = au_colors("red"),
  pval_cutoff = 0.05,
  add_labels = F,
  num_lab_genes = 15,
  label_col = NA,
  lab_size = 3,
  dirs = "y",
  nudge_x1 = 10,
  nudge_x2 = -10,
  nudge_y1 = 0,
  nudge_y2 = 0,
  labsize = NA,
  fillcol = scales::alpha(c("white"), 0),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Volcanoplotter
}
