% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/au_colors.R
\name{au_colors}
\alias{au_colors}
\title{Aarhus University brand colors}
\usage{
au_colors(...)
}
\arguments{
\item{...}{Name(s) of color(s) to be used.}
}
\value{
A named vector with the chosen colors.
}
\description{
Aarhus University brand colors
}
\examples{
au_colors("red", "blue")
}
