% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotter.R
\name{scatterplotter}
\alias{scatterplotter}
\title{Create scatterplots}
\usage{
scatterplotter(
  data,
  x_val,
  y_val,
  col_val = NA,
  style = "light",
  colors = au_colors(),
  y_lab = "y",
  x_lab = "x",
  title = "",
  fit = "none",
  discrete = TRUE,
  linecolor = "black",
  pointcolor = "black",
  corr_method = "pearson",
  alternative = "two.sided",
  fit_method = "glm",
  se = FALSE,
  labels = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Create scatterplots
}
