% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplotter.R
\name{violinplotter}
\alias{violinplotter}
\title{Create violin-boxplots}
\usage{
violinplotter(
  data,
  x_val,
  y_val,
  x_lab = x_val,
  y_lab = y_val,
  title = "",
  filter = F,
  filter_col = NA,
  filter_val = NA,
  comp_vec = NA,
  col_vec = NA,
  scale_labs = ggplot2::waiver(),
  fill_vec = NA,
  col_style = "light",
  fill_style = "light",
  display_n = T,
  violin_alpha = 0.5,
  facet_val = NA,
  ...
)
}
\arguments{
\item{data}{The \code{data.frame} to be used for the visualization.}

\item{x_val}{string, the name of the column to plot on the x axis.}

\item{y_val}{string, the name of the column to plot on the y axis.}

\item{x_lab}{string, the x axis label. Default is the string passed into \code{x_val}.}

\item{y_lab}{string, the y axis label. Default is the string passed into \code{y_val}.}

\item{title}{string, the title of the plot to be displayed on top. Default is \code{""}.}

\item{filter}{boolean, if \code{TRUE}, it allows filtering of \code{data}. Default is \code{FALSE}.}

\item{filter_col}{string, the name of the column to filter if \code{filter = TRUE}. Default is \code{NA}.}

\item{filter_val}{vector, the values to keep if \code{filter = TRUE}.}

\item{comp_vec}{list of vectors containing the comparisons to be passed into \code{stat_compare_means}.
Default is \code{NA}, and in this case, no comparisons are made.}

\item{col_vec}{vector containing the colors to be used for the \code{color} aesthetic. Default is \code{NA}. If unspecified, the function uses \code{au_colors()}.}

\item{scale_labs}{vector, names to display on the x axis ticks. By default, it shows the names as they appear in \code{x_val}.}

\item{fill_vec}{vector containing the colors to be used for the \code{fill} aesthetic. Default is \code{NA}. If unspecified, the function uses \code{au_colors()}.}

\item{col_style}{string, palette style to be used for \code{scale_color_au}. Default is \code{light}. Style is only applied if \code{col_vec} remains \code{NA}.}

\item{fill_style}{string, palette style to be used for \code{scale_fill_au}. Default is \code{light}. Style is only applied if \code{fill_vec} remains \code{NA}.}

\item{display_n}{boolean, if \code{TRUE}, the plot displays the sample size appended to the title. Default is \code{TRUE}.}

\item{violin_alpha}{num, opacity of the violin plot. Default is \code{0.5}.}

\item{facet_val}{string, the name of the column to facet by. Default is \code{NA}.}

\item{...}{other parameters passed into \code{stat_cor()} or \code{facet_wrap()}.}
}
\value{
A ggplot object.
}
\description{
Enables the creation of violin-boxplots in a convenient and customizable manner.
}
\examples{
violinplotter(data = iris, x_val =  "Species", y_val = "Petal.Width",
display_n = FALSE)

violinplotter(data = iris, x_val = "Species", y_val = "Petal.Width",
filter = TRUE, filter_col = "Species", filter_val = c("setosa", "virginica"),
comp_vec = list(c("setosa", "virginica")), title = "Comparing Setosa and Virginica")
}
